`Hanami::RakeTasks` -> `Hanami::CLI::RakeTasks`
